Move syscall struct
